requirements.txt
Flask==2.2.5
mysql-connector-python==9.0.0
Flask-Login==0.6.3
Flask-WTF==1.2.1
gunicorn==23.0.0
